#109: turned Repeat into an envelope
